revised parsing structures
modify path generation to use the path and path/filepath packages in
order to ensure that the import paths we encounter as properly handled
reglardless of the system.

modify parsing to ensure everything is pre-processed before attempting
to generate files.

add the ability to summarize the structures generated, for debugging.