add encoding sniffing and brute-force for decoding.
encoding sniffing is done by allowing Registered Encoders and Decoders
to specify what a potential first rune is.  This is primarily useful
for string based encodings where the first rune is conistent.  Some
examples of this would be XML with '<', and JSON with '{', or '['.

brute-force encoding determination will attempt to use all encodings
that have not yet failed, and Decode with each of them.  If any
succeed, that is assumed to be the encoding.  This has some potential
problems, firstly, there are encoding overlaps for different mime
types, such as tag languages like HTML and XML.  They would likely be
identified as the same encoding types.  Secondly, this is horribly
slow, and is only used as a last resort.

Please keep in mind that the
best thing to do is to specify your Content-Type, and Accept Headers.
As this will give the Default Encoder all the information needed to
resolve the encoding on it's own, without having to fall back on
sniffing, or brute force methods.

add encoding sniffing for Decoding methods

add brute-force encoding detection for Decoding methods

add several tests to ensure that encoding detection is performing well
in several circumstances.