modify generated templates for various reasons
- The template generation has been modified to take special
consideration for golang.org/x/net/context.Context. If you write an
interface that uses this item as a parameter for a method, and it is
imported as "golang.org/x/net/context", it will identify this, and
will pass-through the context generated from a request to it.

- I've modified the Import type stored within Package and File
to be pointers so that they can be mutated as new information is
determined about them.

- Various imports have been marked as a Parameter if it is used as
either a parameter or a return result of a Method within an Interface,
or as an Embeded type, if it is Embeded within an Interface.  This has
been done to make filtering out Imports more precise.

- New fields have been added to the Template structures for Param, to
identify whether the field is a Context or not, and to the Method
Template to specify whether this methods has context in it's
parameters, and it's name.  The name from this template is used to
name the Context regardless of whether it is a parameter or not. It
is defaulted to "_ctx".

- Add DefaultRequestTimeout so that the user can specify what they
would like the timeout to be overall, instead of having it hard-coded
to time.Second. As it stands, it currently defaults to time.Second.

- Some logic cleanup in the template structures as things became a bit
more simple.