add the ability to encode errors of various types
Errors can be transmitted across from the server in a more consistent
manner now. One can choose to use the Default Encoder's
ResponseEncoder as the github.com/go-kit/kit/transport/http.Server's
errorEncoder, and it will do it's damnedest to ensure that an error
of some sort is transmitted across the line.

- It achieves this by wrapping the underlying error with information
that helps it to give it's best effort in decoding. It does this by
storing two different representations of the error, along with the
type of the error in a wrapped error structure.

- In order to ensure that the encoders and decoders have the ability
to encode/decode the underlying errors, the error type itself must be
registered with the github.com/ayiga/go-kit-middlewarer/encoding
package's RegisterError function. Please note, that unexported types
with unexported fields are considered unable to be encoded or decoded
so they should be avoided.  To this end, errors created from
errors.New, are not encodable, so the summary will come through, but
you will be unable to get and equality test against them.

The ResponseDecoder now checks the status code to determine whether it
should be looking for an error, or not. In general, if the response
code is not 2xx, then it is considered an error being transmitted,
instead of the response type. If the content-type is text/plain, then
it is assumed that the default error encoder is used, and the error
will be returned as errors.New. However, if the content-type is
anything else, the assumption is made that the type is the Wrapped
Error being transmitted, and it will attempt to decode that instead.