add extra response error checks in the default handler
this was done because the other case is that it already has a content
type.  It is very unlikely to get to the JSON, XML, or Gob decoders
without a known content-type.  As such, the default handler is the
best place for it.  However, the Encoders and Decoders are able to
be instanciated manually, and as such, I've kept the error handling
code in that location as well.